refactor: move Dual<T> into combine.rs and implement via #[derive(OpPriv)]
Previously Dual<T> lived in a separate dual.rs and all trait impls
(Construction, Semigroup, Monoid, Commutative) were written by hand.

Now it uses #[derive(OpPriv)] for the boilerplate — the same pattern
as Sum, Coalesce, and other ops — and only Op<T>::lift_op_assign is
written manually (the unsafe ptr::read/write reversed operation).

The monoid_where / commutative_where op-attributes handle the
conditional Monoid and Commutative impls without extra hand-written
code. Dual<T> is co-located with CombineIterator in combine.rs since
it is the canonical type backing rfold_final / rreduce.
